refactor: move project files to an api sub-directory
note:
* this makes room for processes and their phases
